Reading large assumptions file modified by tellg() and seekg()

+++ src_mpi/mpi_base.h
@@ -98,7 +98,7 @@ public:
 	bool IsSolveAll;
 	double max_solving_time; // max time in seconds for solving particular SAT subproblem
 	unsigned keybit_count;
-	unsigned assumptions_count;
+	unsigned long long assumptions_count;
 	string rslos_table_name;
 	int max_nof_restarts;
 	char *input_cnf_name;
